Implemented basic osipi_fit method to the base class
My thinking is that this is the default function for fitting. It basically passes arguements to the ivim_fit method of standardized version of the algorithm submission. But before a fit is performed, checks should be made to make sure algorithm requirements are fulfilled. This is yet to be implemented